+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Optimize the selection style of items in the menu.
 - Optimize code block styles.
+- Optimize some element styles on Issue/PR page
 
 ##### Template File
 
@@ -15,3 +16,5 @@
 
 - Fixed extra lines under heatmap.
 - Supplement the PR operation panel status style in PR.
+- Fixed avatar size for approving users in PR
+- Fixed bot label styles in Issue/PR